source-systems/requirements.txt
boto3>=1.28.0
pandas>=2.0.0
psycopg2-binary>=2.9.9
pyarrow>=14.0.0

